refactor: 메소드명 변경, login 패키지 추가
