--- src/main/java/auth/UserDetailsFactory.java
@@ -1,8 +1,0 @@
-package auth;
-
-public interface UserDetailsFactory {
-    UserDetails createUserDetails(String username, String password);
-
-    UserDetails createUserDetails(Long principal, String role);
-
-}
