+++ src/main/java/auth/login/LoginController.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.login;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
