+++ src/main/java/auth/login/TokenRequest.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.login;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class TokenResponse {
-    private String accessToken;
+public class TokenRequest {
+    private String username;
+    private String password;
 }
