+++ src/main/java/auth/login/TokenResponse.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.login;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class TokenRequest {
-    private String username;
-    private String password;
+public class TokenResponse {
+    private String accessToken;
 }
