+++ src/main/java/auth/login/UserDetails.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.login;
 
 
 import lombok.AllArgsConstructor;
