+++ src/main/java/auth/login/UserDetailsFactory.java
@@ -1,0 +1,8 @@
+package auth.login;
+
+public interface UserDetailsFactory {
+    UserDetails makeUserDetails(String username, String password);
+
+    UserDetails convertToUserDetails(Long id, String role);
+
+}
