+++ src/main/java/nextstep/WebMvcConfiguration.java
@@ -1,6 +1,9 @@
 package nextstep;
 
 import auth.*;
+import auth.login.LoginController;
+import auth.login.LoginService;
+import auth.login.UserDetailsFactory;
 import lombok.AllArgsConstructor;
 import nextstep.member.MemberDao;
 import nextstep.member.UserDetailsFactoryImpl;
